+++ packages/plugin/src/rspack.ts
@@ -1,0 +1,4 @@
+import { createRspackPlugin } from "unplugin";
+import { unpluginFactory } from ".";
+
+export default createRspackPlugin(unpluginFactory);